# Conflicts:
#	Cargo.lock
#	crates/jolt-field/src/akita.rs
#	crates/jolt-field/src/transcript_challenge.rs
#	crates/jolt-verifier-derive/src/lib.rs
#	crates/jolt-verifier/src/verifier.rs
#	crates/jolt-witness/src/field_inline/mod.rs

+++ tracer/src/instruction/virtual_assert_word_alignment.rs
@@ -18,7 +18,7 @@ impl VirtualAssertWordAlignment {
         cpu: &mut Cpu,
         _: &mut <VirtualAssertWordAlignment as RISCVInstruction>::RAMAccess,
     ) {
-        let address = cpu.x[self.operands.rs1 as usize].wrapping_add(self.operands.imm);
+        let address = cpu.x[self.operands.rs1 as usize].wrapping_add(self.operands.imm as i64);
         assert!(
             address & 3 == 0,
             "RAM access (LW or LWU) is not word aligned: {address:x}"
